Update patch set 1
Patch Set 1:

> Hi Frank, Apologies for the inconvenience caused by the commit I
 > made. I am new to NFS community and was using my repository for the
 > first time. My intention was to save my work in my repository, but
 > didn't know committing using gerrithub would lead to unnecessary
 > notifications.
 > Malahal has suggested to use github for committing work to my
 > repository. I will follow his suggestion.
 > Thanks.

No problem, and sorry if I came off cranky... It's not really clear to folks that they only need a gerrithub repo if they are going to post code for review and do merges themselves, however, once you have code you want to solicit review of any sort for, please submit to ffilz/nfs-ganesha gerrithub. You can put a temporary WIP in the patch title (to be removed later with gitrebase -i [reword] or git commit --amend) to let me know the patch is not yet ready for merging.

That way, all the comment history is preserved which may be valuable sometime in the future.

Patch-set: 1