Remove NOENT return code from mdcache_readdir
Change backported from upstream commit
48f8de3626bfdf7ba7c64b38f60755ef78de2abc
The current code sets status to NOENT when
EOD is reached. Removed the return code as NFS3ERR_NOENT
is not listed in the ERRORS list of READDIRPLUS3 Procedure in the RFC.
Also the RFC doesn't clearly mention that an error should be returned
when EOD is reached.

Change-Id: Iba6f5905cbd962acb9aafa8620797345da78013a
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>