release file sys after file sys is unclaimed
currently in GPFS after last export of a file system is removed,
the file system is unclaimed but not removed. If a file system is deleted
and a new file system is added which has the same dev major, minor numbers as the
previous file system then it is skipped as duplicate file system.
Made changes by calling "release_posix_file_system" after the file system is unclaimed.

Change-Id: If3b73aef1ea7b595143d3cdd7507caaeee09079c
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>