release file sys after file sys is unclaimed
currently in GPFS after last export of a file system is removed,
the file system is unclaimed but not removed. If a file system is deleted
and a new file system is added which has the same dev major, minor numbers as the
previous file system then it is skipped as duplicate file system.
Made changes by calling "release_posix_file_system" after the file system is unclaimed.

Change-Id: Ic9d69f41ba15913c216278373871e8fd5ea1af88
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>