NLM: Use caller name when SM_NOTIFY sent by localhost
Currently when SM_NOTIFY request is sent by a NFS client,
the requesting client IP is used and we release locks
for that client IP. This works fine as the caller name
in the SM_NOTIFY request corresponds to the name of the NFS client
and the IP sending the request is the client IP.
But if there was a NFS client which had acquired locks and if
the client is dead, the server can release locks for that client by
simulating the SM_NOTIFY request for the client by forming a
SM_NOTIFY request with the caller name of the client. But with
current implementation this case doesn't get handled as the current
implementation checks if locks were acquired by the IP which has
sent the SM_NOTIFY request.
To fix this we now check if the requesting IP is for 'localhost'
and the IP for the caller name in SM_NOTIFY doesn't match
with IP for 'localhost' then use the IP for the caller name as the
requesting IP.

Change-Id: I4315ace60987b277f5b582d72d37531c584c61e8
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>