Wakeup LRU thread if fds_hard_limit/fds_hiwat reached
Currently we increment/decrement the FD count after a
file is opened/closed, but before opening a file there
is NO check for the current FD count to compare it
against the "lru_state.fds_hard_limit" or
"lru_state.fds_hiwat" (FD high water mark).
With this we can run out of FDs if
"number of FDs opened" > "FD limit" within the
LRU_Run_Interval (default 90 seconds),
To fix this we now call "mdcache_lru_fds_available()"
before opening a file, so that if the FD limit is reached
we wake up the LRU thread to free some FDs to make space
for new FDs.

Change-Id: I5bb8ad4bd0b11afa8bebd744c7f7d80bcbb148aa
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>