Use 'v6disabled' flag to know if IPv6 is disabled
Currently ganesha startup on a system with IPv6 disabled fails
with message "Register_program :DISP :FATAL :Cannot register
NFS V3 on UDPv6". This is because there is no check to
know if IPv6 is enabled or not in Register_Program to decide if
a protocol needs to be registered for IPv6 or not.
Fixed this by checking 'v6disabled' flag to know if IPv6 is
disabled. If yes, in Register_Program we do UDP_REGISTER/TCP_REGISTER
for netconfig_udp4/netconfig_tcp4. Else we check if
netconfig is enabled for udp6/tcp6. If no in Register_Program
we do UDP_REGISTER/TCP_REGISTER for netconfig_udp4/netconfig_tcp4,
else in Register_Program we do UDP_REGISTER/TCP_REGISTER
for netconfig_udp6/netconfig_tcp6.

Change-Id: I71b9f8d96c5d9d18b435fd8afedc689720e6d34e
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>