Handle get_port() return value in nfs_rpc_execute()
Currently in nfs_rpc_execute(), if a transports 'ss_family'
is not AF_INET/AF_INET6, the get_port() returns -1.
But we ignore this return value in nfs_rpc_execute() and
call get_gsh_client() which will assert() as 'ss_family'
is not AF_INET/AF_INET6.
Changed nfs_rpc_execute() to check the return value of
get_port(), if it is -1 we respond to the request with svcerr
and don't call get_gsh_client().

Change-Id: I9bb7baa3cb0411e9b11e910617ea018970048344
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>