Repopulate chunk after reacquiring a dropped lock
In mdcache_readdir_chunked() when first_pass=false, it
means we have a valid chunk. But in the next pass if
write lock has not been already acquired then we drop
the existing lock and acquire a write lock. Here after
dropping the lock there is a possibility that another
thread may acquire the lock and discard the chunk.
Thus, we not cannot trust the chunk pointer and we
should repopulate the chunk.

Change-Id: I41c806b96e6cd26789693bf5cfdb070675f0ffac
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>