MDCACHE - Keep an entry pointer in dirent when ref'd
When a dirent has a ref on an entry, it may still not be able to look
the entry up, because it may become unreachable.  In this case, we can't
drop the ref.

Instead, keep a temporary pointer to the entry that is valid only when
we have the ref, and use this instead of looking it up.

Change-Id: I1b87d4fb4201885dd29220dd5515fdaea39ff7c4
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit f682cd9d424dd1f6d00811be5ab605b5f1811557)
(cherry picked from commit b31d7d2075901c06edc837d39466829a8bb74d49)