If op_ctx not set during shutdown then log debug trace for an entry and not assert
Currently during shutdown, while destroying FSALs if an extra handle is hanging
around for MDCACHE then mdache_lru_clean() is called. In mdcache_lru_clean() if
op_ctx is not set then we may assert(). The idea to have this assert() was to detect
refcnt leaks, if there are no refcnt leaks then mdcache_lru_clean() wouldn't get
called when destroy_fsals() is executing. But in case of refcnt leaks
mdcache_lru_clean() gets invoked when destroying FSALs and if op_ctx==NULL then we
assert().
This is a temporary fix, to avoid assert() when op_ctx=NULL during shutdown in
mdcache_lru_clean(). With this change we don't assert() if shutdown is going on,
but instead log a debug trace to tell about the refcount leak for an entry.
Note: This fix will be removed when we have the fix for refcnt leak.

Change-Id: I7b7b03fc7344395122ade1761d62ec261444c1f6
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>