mdcache_create_handle: Don't lookup for parent for a root object
Currently in mdcache_create_handle() after getting an object  we
call mdc_get_parent() for the entry. But if the object  is a
root object we would be doing lookup of .. of the root, which
doesn't look correct.
To fix this prior to call mdc_get_parent() check if the object is
a root object, if yes then don't call mdc_get_parent() for it.

Change-Id: Ibf25fd283573c9fa28ae0452d96f00f294f66828
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>